Make ESStoreWrapper class public
We want to access it in subclasses of ESDepotStore.
