+++ winterwell.depot/src/com/winterwell/depot/ESDepotStore.java
@@ -220,17 +220,3 @@ public class ESDepotStore implements IStore {
 		// throw new TodoException();
 	}
 }
-
-class ESStoreWrapper {
-
-	private transient Object artifact;
-
-	public ESStoreWrapper(Object artifact) {
-		this.artifact = artifact;
-		// Base 64 encode too??
-		this.raw = XStreamUtils.serialiseToXml(artifact);
-	}
-
-	public String raw;
-	
-}
